requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0
django-tinymce==3.4.0
gunicorn==20.1.0
Pillow==8.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
PyMySQL==1.0.2
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
whitenoise==5.3.0
